+++ config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
   scope module: :avocado do
+    resources :affirmations, only: %i[new show create]
     resources :recoveries, only: %i[new create edit update]
     resources :registrations, only: %i[new create]
     resources :sessions, only: %i[index new create destroy]